physics_guided_simulation/requirements.txt
# Core Python k-Wave Ultrasound Simulation Pipeline Dependencies

# k-Wave library
python-kwave>=0.3.0

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Data handling
pandas>=1.3.0

# Optional GPU acceleration (uncomment if using GPU)
# cupy>=10.0.0

# Development and utilities
tqdm>=4.62.0
jupyter>=1.0.0

# For MATLAB file compatibility
h5py>=3.0.0

# Type hints (for Python < 3.9)
typing-extensions>=4.0.0

# Optional visualization
seaborn>=0.11.0

# For progress tracking
psutil>=5.8.0 